+++ qutip/tests/core/test_ptrace.py
@@ -136,9 +136,12 @@ def test_ptrace_operator(dtype, sel):
                              ([5, 2, 3], [2, 1]),
                              ([5, 2, 3], [0, 2]),
                              ([5, 2, 3], [0, 1]),
-                             ([2]*6, [3, 2]),
+                             ([5, 2, 3], [0]),
+                             ([2]*6, []),
+                             ([2]*6, [3]),
                              ([2]*6, [0, 2]),
-                             ([2]*6, [0, 1]),
+                             ([2]*6, [0, 1, 4]),
+                             ([2]*6, [0, 1, 2, 3, 4, 5]),
                          ])
 def test_ptrace_ket_specialization_matches_old_implementation(dtype, dims, sel):
     """Kets have a different implementation for ptrace.
